requirements.txt
# Django Web packages
django
djangorestframework
markdown
django-filter
django-cors-headers
django_daraja
python-dotenv

# Twilo package
twilio

# AI packages
llama-index
pypdf
llama-index-llms-groq
tensorflow